Test META markup after removing transitions
Currently the META injections are not updating fast enough to be picked up by Google, so this means that there is an issue with the data not being loaded correctly on page load.  If this does not address the issue, we will have to look into SSR, as that means that the AWS stuff is using AJAX, and that just won't work if we want SEO friendly pages.